+++ src/com/org/gnos/db/model/WasteMiningCost.java
@@ -7,7 +7,7 @@ public class WasteMiningCost extends FixedOpexCost{
 	public WasteMiningCost(){
 		super();
 		this.id = -1;
-		this.costData = new LinkedHashMap<Integer, Integer>();
+		this.costData = new LinkedHashMap<Integer, Float>();
 	}
 	
 }
